refactor(dashboard): extract shared files logic to separate component
Move shared files rendering logic from page component to dedicated SharedContent component
Add suspense boundary with loading skeleton for better user experience
