+++ app/dashboard/shared/page.tsx
@@ -1,9 +1,9 @@
 import { Metadata } from "next";
-import DashboardLayout from "../../../components/dashboard-layout";
-import FileGrid from "../../../components/file-grid";
-import { getEmailFromUserId } from "../../../lib/getEmailFromUserId";
-import { dbAdmin } from "../../../lib/firebase-admin";
+import { Suspense } from "react";
 import { headers } from "next/headers";
+import DashboardLayout from "../../../components/dashboard-layout";
+import FileSkeleton from "../../../components/FileSkeleton";
+import SharedContent from "../../../components/pages/shared";
 
 export const metadata: Metadata = {
   title: "StellarSync | Shared",
@@ -19,7 +19,6 @@ export default async function SharedPage() {
   if (!userId) {
     return (
       <DashboardLayout>
-        <p className="p-6"></p>
         <div className="col-span-full flex h-40 flex-col items-center justify-center rounded-lg border border-dashed p-4 text-center">
           <p className="text-lg font-medium">Not signed in.</p>
         </div>
@@ -27,31 +26,11 @@ export default async function SharedPage() {
     );
   }
 
-  const email = await getEmailFromUserId(userId);
-
-  let files: Array<any> = [];
-  const base = dbAdmin
-    .collection("uploadedFiles")
-    .where("userEmail", "==", email);
-
-  const snap = await base.get();
-  snap.docs.forEach((d) => files.push({ id: d.id, ...d.data() }));
-
   return (
     <DashboardLayout>
-      <div className="flex flex-col gap-6">
-        <div className="flex flex-col gap-2">
-          <h1 className="text-2xl font-bold tracking-tight">Shared Files</h1>
-          <p className="text-muted-foreground">
-            Manage files you've shared with others.
-          </p>
-        </div>
-
-        <FileGrid
-          fileList={files.filter((file) => file.shared)}
-          view="shared"
-        />
-      </div>
+      <Suspense fallback={<FileSkeleton />}>
+        <SharedContent userId={userId} />
+      </Suspense>
     </DashboardLayout>
   );
 }
